+++ familyhelper-finance-stack/src/main/java/com/dwarfeng/familyhelper/finance/stack/bean/entity/Poab.java
@@ -11,6 +11,8 @@ import com.dwarfeng.subgrade.stack.bean.entity.Entity;
  */
 public class Poab implements Entity<PoabKey> {
 
+    public static final int PERMISSION_LEVEL_OWNER = 0;
+
     private static final long serialVersionUID = 6405468025576702046L;
 
     private PoabKey key;