+++ familyhelper-finance-stack/src/main/java/com/dwarfeng/familyhelper/finance/stack/exception/AccountBookNotExistsException.java
@@ -1,0 +1,40 @@
+package com.dwarfeng.familyhelper.finance.stack.exception;
+
+import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+
+/**
+ * 账本不存在异常。
+ *
+ * @author DwArFeng
+ * @since 1.1.0
+ */
+public class AccountBookNotExistsException extends BalanceOperateException {
+
+    private static final long serialVersionUID = 4394311960147137564L;
+
+    private final LongIdKey accountBookKey;
+
+    public AccountBookNotExistsException(LongIdKey accountBookKey) {
+        this.accountBookKey = accountBookKey;
+    }
+
+    public AccountBookNotExistsException(String message, LongIdKey accountBookKey) {
+        super(message);
+        this.accountBookKey = accountBookKey;
+    }
+
+    public AccountBookNotExistsException(String message, Throwable cause, LongIdKey accountBookKey) {
+        super(message, cause);
+        this.accountBookKey = accountBookKey;
+    }
+
+    public AccountBookNotExistsException(Throwable cause, LongIdKey accountBookKey) {
+        super(cause);
+        this.accountBookKey = accountBookKey;
+    }
+
+    @Override
+    public String getMessage() {
+        return "账本 " + accountBookKey + " 不存在";
+    }
+}