+++ CHANGELOG.md
@@ -16,6 +16,8 @@
   - com.dwarfeng.familyhelper.finance.stack.bean.entity.User。
   - com.dwarfeng.familyhelper.finance.stack.bean.entity.TotalBalanceHistory。
   - com.dwarfeng.familyhelper.finance.stack.bean.entity.BankCardBalanceHistory。
+- 创建余额操作服务。
+  - com.dwarfeng.familyhelper.finance.stack.service.BalanceOperateService。
 
 #### Bug修复